requirements.txt
kafka-python
pyspark
pymongo
streamlit
pandas
